pyproject.toml
[tool.poetry]
name = "rhubarb"
version = "0.0.1"
description = "A Python framework for multi-modal document understanding with Amazon Bedrock"
authors = ["Rhubarb Developers <rhubarb-developers@amazon.com>"]
license = "Apache 2.0"
readme = "README.md"
homepage = "https://github.com/awslabs/rhubarb"
repository = "https://github.com/awslabs/rhubarb"
documentation = "https://github.com/awslabs/rhubarb"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
packages = [
    { include = "rhubarb", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9"
pillow = "^10.3.0"
pdfplumber = "^0.11.0"
boto3 = "^1.34.71"
botocore = "^1.34.71"
jsonschema = "^4.21.1"
pydantic = "^2.6.4"
pyarrow = "^15.0.2"
numpy = "^1.26.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
pre-commit = "^3.7.0"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
length-sort = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Homepage" = "https://github.com/awslabs/rhubarb"
"Issues" = "https://github.com/awslabs/rhubarb/issues"
